Add files via upload
The purpose of this project is to comparatively analyze the effectiveness of prediction algorithms on stock market data and get a general insight into this data through visualization to predict future stock behavior and value at risk for each stock. The project encompasses the concept of Data Mining and Statistics. This project makes heavy use of NumPy, Pandas, and Data Visualization Libraries.